Assistant checkpoint: Add babel plugin and update OpenAI config
Assistant generated file changes:
- src/components/LandingPage.js: Update OpenAI configuration to allow browser usage

---

User prompt:

it says this (but @babel/plugin-proposal-private-property-in-object is not an option):

One of your dependencies, babel-preset-react-app, is importing the
"@babel/plugin-proposal-private-property-in-object" package without
declaring it in its dependencies. This is currently working because
"@babel/plugin-proposal-private-property-in-object" is already in your
node_modules folder for unrelated reasons, but it may break at any time.

babel-preset-react-app is part of the create-react-app project, which
is not maintianed anymore. It is thus unlikely that this bug will
ever be fixed. Add "@babel/plugin-proposal-private-property-in-object" to
your devDependencies to work around this error. This will make this message
go away.